requirements.txt
torch
torchvision
kaggle
pandas
numpy
matplotlib
seaborn
pillow
scikit-learn
optuna